requirements.txt
Flask==3.0.2
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7 
Flask-SQLAlchemy==3.0.5
PyMySQL==1.0.3
python-dotenv==1.0.0
cryptography==40.0.1
